Day 20
This could be a lot better, but only takes a couple of seconds.
Simulation isn't needed for part 1 (acceleration dominates), but
I didn't like the idea of doing that sort. Could also order by
distance for part 2, and only compare things at same distance, but
again, sorting is hampered by the fact that I picked a hashmap...